Updating delete post confirmation

+++ admin/includes/view_all_posts.php
@@ -129,7 +129,7 @@ if (isset($_POST["checkBoxArray"])) {
             echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
 
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onClick=\"javascript: return confirm('Ar you sure you want to delete');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
             echo "</tr>";
         }
     ?>
